+++ uoft_mcp/acorn/__init__.py
@@ -1,0 +1,1 @@
+"""ACORN integration namespace; authentication is currently the supported capability."""